requirement.txt
asgiref==3.4.1
attrs==21.2.0
Babel==2.9.1
cached-property==1.5.2
certifi==2021.5.30
charset-normalizer==2.0.6
Django==3.2.7
django-colorfield==0.4.2
django-money==2.0.3
django-phonenumber-field==5.2.0
django-render-partial==0.4
django-widget-tweaks==1.4.8
idna==3.2
isodate==0.6.0
lxml==4.6.3
phonenumbers==8.12.31
Pillow==8.3.2
platformdirs==2.3.0
py-moneyed==1.2
pytz==2021.1
requests==2.26.0
requests-file==1.5.1
requests-toolbelt==0.9.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
zeep==4.1.0
